RH5: vmbus: fix array indexing error
In channel.c, function vmbus_get_debug_info(), when the following line execute, the system panics:
    debuginfo->clientmonitor_connectionid = monitorpage->parameter[monitor_group][monitor_offset].connectionid.u.id;
The array indexes are computed as:
    u8 monitor_group = (u8)channel->offermsg.monitorid / 32;
    u8 monitor_offset = (u8)channel->offermsg.monitorid % 32;
When the channel->offermsg.monitorid = 255, monitor_group = 7
In hyperv_vmbus.h, the hv_monitor_page structure defines the parameter variable as:
    struct hv_monitor_parameter parameter[4][32];
monitorid has a range of 0 - 255. Dividing by 32 results in a range of 0 - 7 for monitor_group, which exceeds the array size of 4.