hv_netvsc: fix a UDP RSS hashing bug for rhel 6.x tree
In the 32 bits of CentOS 6.7, the skb structure for DHCP
Discovery
message may have transport_header point to NULL.
In this case,
udp_hdr(skb) will return NULL reference, and current
implementation of
udp_hdr(skb)->source will crash the kernel when
bootup it with a NIC set
to dhcp mode.
This patch will check the udp_hdr(skb) before we use it.