Drivers: hv: vmbus: On write cleanup the logic to interrupt the host
Signal the host when we determine the host is to be signaled.
The currrent code determines the need to signal in the ringbuffer code
and actually issues the signal elsewhere. This can result in the host
viewing this interrupt as spurious since the host may also poll the
channel. Make the necessary adjustments.