[hv-rhel6.x] net: make ndo_get_stats64 a void function
Author: stephen hemminger <stephen@networkplumber.org>
commit bc1f44709cf27fb2a5766cadafe7e2ad5e9cb221

The network device operation for reading statistics is only called
in one place, and it ignores the return value. Having a structure
return value is potentially confusing because some future driver could
incorrectly assume that the return value was used.

Fix all drivers with ndo_get_stats64 to have a void function.