RH6: Add timesync_mode parameter to control TimeSync behavior
The TimeSync IC was setting the local clock with every sample from the
host. This can interfere with NTP's gradual slewing of the guest clock.

Add a module parameter to control how often the clock should be
disciplined. The default setting tells TimeSync to update the clock only
during startup. Setting this param to '1' tells TimeSync to update
the clock on every sample, and tracks the host clock more closely.