rh7: vmbus: don't invoke tasklet on wrong CPU
During initialization, the channel initialization code schedules the tasklet
to scan the VMBUS receive event page (simulates an interrupt). The problem was
that it invokes the tasklet on a different CPU from where it normally runs
and therefore if an event is present, it will clear the bit but not find
the associated channel.

This can lead to missed events, typically stuck tasks, during bootup
when sub channels are being initialized.

Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>