Tools: hv: vss: Thaw the filesystem and continue if freeze call has timed out
6113e3d28149dee33f10e92df6a16efb2b9cf473

If a FREEZE operation takes too long, the driver may time out and move
on
to another  operation. The daemon is unaware of this and attempts to
notify the driver that the FREEZE succeeded. This results in an error
from
the driver and the daemon leaves the filesystem in frozen state.

Fix this by thawing the filesystem and continuing.