RH6: Add disable_mtrr_trim to kernel cmdline upon install
This works around issue #321. Summarized here:

Old RHEL 6.x kernel has a bug, which can be exposed when the VM is
running on WS 2016, where the reported max supported physical address
has 44 bits. On WS 2012 R2, the max physical address has only 42 bits,
and the bug is not triggered.

The old kernel needs to pick the patch:
x86: Fix /proc/mtrr with base/size more than 44bits
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=d5c78673b1b28467354c2c30c3d4f003666ff385

or, we can use the "disable_mtrr_trim" kernel parameter to work around the issue.
There is not perf issue with this, because Hyper-V always ignores
the MTRR setting for conventional RAM of a VM).