RH7: netvsc: Save the previous section size before reinitializing buffer
This fixes an issue where we would reset the section size upon
tearing down and re-adding a net_device for an MTU change. If we
continuously increase the MTU, this leads to a decrease in section
count while maintaining the same section size. Since buffer size
is calculated as section_count*section_size, we eventually get a
use very small buffer size that is rejected by the host.