RH6: pci-hyperv: support sr-iov
Hi Simon,

Could you run a round of tests for this patch for accelerated networking
on RHEL6 please? I haven't done much beyond see that it boots and that
the network card is initialized.

======

Based on Vitaly's work.

Relevant dmesg output on boot:

	hv_vmbus: registering driver hv_pci
	hv_pci 0c069445-8b3e-4aa6-a012-c18a1681378e: PCI host bridge to bus a012:00
	pci_bus a012:00: root bus resource [mem 0xfe0000000-0xfe07fffff window]
	pci a012:00:02.0: BAR 2: assigned [mem 0xfe0000000-0xfe07fffff 64bit pref]
	mlx4_core: Mellanox ConnectX core driver v2.2-1 (Feb, 2014)
	mlx4_core: Initializing a012:00:02.0
	mlx4_core a012:00:02.0: Detected virtual function - running in slave mode
	mlx4_core a012:00:02.0: Sending reset
	mlx4_core a012:00:02.0: Sending vhcr0
	mlx4_core a012:00:02.0: HCA minimum page size:512
	mlx4_core a012:00:02.0: Timestamping is not supported in slave mode

Ugly hacks to get this running without redhat's support:

 - Export few internal kernel functions using kallsyms_on_each_symbol().
 - Hook arch_teardown_msi_irqs() and msi_compose_msg() directly using
simple trampolines.
 - Set IRQ chip affinity function directly.

I've left a few debug printk() in for now to make testing simpler, but
they should be removed before merging this in.

Signed-off-by: Sasha Levin <alexander.levin@microsoft.com>