RH6: hv_netvsc: Use vf_xmit when not in softirq
When netvsc_start_xmit() is called from softirq, there is a possible
race in VF TX path, which may cause skb->dev to be NULL.
This patch works around the issue by only calling vf_xmit when not in
softirq, otherwise falling back to synthetic path.
Perf results remain similar to that before this patch.

Signed-off-by: Haiyang Zhang <haiyangz@microsoft.com>